temp for grannou

+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: mgamil <mgamil@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/25 21:25:30 by mgamil            #+#    #+#             */
-/*   Updated: 2022/12/05 20:53:22 by mgamil           ###   ########.fr       */
+/*   Updated: 2022/12/07 22:47:58 by mgamil           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,7 +40,7 @@ int					ft_isdigit(int c);
 int					ft_isprint(int c);
 
 char				*ft_itoa(int n);
-int					ft_atoi(const char *str);
+long long			ft_atoi(const char *str);
 int					ft_tolower(int c);
 int					ft_toupper(int c);
 void				ft_swap(void *a, void *b, int size);
